fix!: remove forced datasets filter on tags in CreateScenarioButton
BREAKING CHANGE: the list of datasets provided to the CreateScenarioButton
component will no longer be filtered, showing the whole list of datasets
when creating a new scenario, including the datasets that don't have the
tag "dataset". If your webapp still needs this behavior, add the filter
mechanism before forwarding the list of datasets to the props of
CreateScenarioButton.