fix: [PROD-12944] fix Table crash when showing large files in fullscreen
Multiple re-renders of the Table seemed to be called because of some `useMemo`
dependencies changing at each render. This caused a lock of the grid component
("cannot get grid to draw rows when it is in the middle of drawing rows").
In particular, the `AgGridReact` props `columnDefs`, `columnTypes` and
`defaultColDef` caused the bug in our case.