fix: fix Table hidden by errors panel
In scenario view, tables could be hidden when an errors panel was
visible. This commits adds a minimum height for the table and a
scrollbar to make it always visible.
Other minor changes have been added to keep the correct height in
fullscreen and when a '100%' height is provided in the component
props (e.g. in the 'dataset details' dialog).

(cherry picked from commit 85efaa232578e2a6266642d69351b175c63156bf)