feat: [PROD-14225] make attributes order configurable in CytoViz component
This commit adds an 'elementsMetadata' prop, that can be used to define the
desired order of attributes when selecting a node or edge in the visualization
tool. The expected format for this prop is:
   {
     attributesOrder: {
       nodes: { nodeType1: ['someImportantAttribute', 'attribute1', 'attribute2'] },
       edges: { edgeType1: ['attributeA', 'attributeB']}},
     }
   }

If some attributes are missing in the list, they will be displayed after all the
other attributes.